refactor: eliminate duplicate code and clean up unused files
