+++ apps/backend/src/app.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
 import { UsersModule } from './users/users.module';
 import { AuthModule } from './auth/auth.module';
@@ -27,6 +26,5 @@ import { PdfModule } from './pdf/pdf.module';
     PdfModule
   ],
   controllers: [AppController],
-  providers: [AppService],
 })
 export class AppModule {}
