+++ apps/backend/src/audit/audit.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Post, Body, UseGuards, Request, Get } from '@nestjs/common';
 import { AuditService } from './audit.service';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
-import { TokenDto } from './dto/token.dto';
+import { TokenDto } from 'src/common/dto/token.dto';
 
 @Controller('audit')
 @UseGuards(JwtAuthGuard)
