+++ apps/backend/src/auth/auth.service.ts
@@ -5,11 +5,11 @@ import {
 } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { HashingServiceProtocol } from './hash/hashing.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { LoginDto } from './dto/login.dto';
 import jwtConfig from './config/jwt.config';
 import { ConfigType } from '@nestjs/config';
-import { UserDto } from './dto/user.dto';
+import { UserDto } from 'src/common/dto/user.dto';
 
 @Injectable()
 export class AuthService {
