+++ apps/backend/src/session/session.controller.ts
@@ -3,7 +3,7 @@ import { SessionService } from './session.service';
 import { CreateSessionDto } from './dto/create-session.dto';
 import { UpdateSessionDto } from './dto/update-session.dto';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
-import { TokenDto } from './dto/token.dto';
+import { TokenDto } from 'src/common/dto/token.dto';
 import { UUID } from 'node:crypto';
 
 @Controller('session')
